mid-seeding

+++ E.P.C/Models/CPU.cs
@@ -6,8 +6,8 @@
         public string Model { get; set; } = string.Empty;
         public int Cores { get; set; }
         public int Threads { get; set; }
-        public float BaseClock { get; set; } // in GHz
-        public float BoostClock { get; set; } // in GHz
+        public double BaseClock { get; set; } // in GHz
+        public double BoostClock { get; set; } // in GHz
         public string SocketType { get; set; } = string.Empty;
         public bool IntegratedGraphics { get; set; }
     }
