+++ E.P.C/Models/CPUFan.cs
@@ -6,7 +6,7 @@
         public string Model { get; set; } = string.Empty;
         public int Size { get; set; } // in mm
         public int RPM { get; set; } // Revolutions per minute
-        public int Airflow { get; set; } // in CFM (Cubic Feet per Minute)
+        public double Airflow { get; set; } // in CFM (Cubic Feet per Minute)
         public int NoiseLevel { get; set; } // in dBA
         public string ConnectorType { get; set; } = string.Empty; // e.g., 3-pin, 4-pin PWM
         public string SupportedSockets { get; set; } = string.Empty; // e.g., LGA1200, AM4
